+++ src/main/java/com/hbm/dim/orbit/SkyProviderOrbit.java
@@ -41,12 +41,15 @@ public class SkyProviderOrbit extends SkyProviderCelestial {
 
 		OpenGlHelper.glBlendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA, GL11.GL_ONE, GL11.GL_ZERO);
 
-		float solarAngle = getCelestialAngle(world, partialTicks, station);
+		if(provider.metrics == null) provider.updateSky(partialTicks);
+
+		float solarAngle = getCelestialAngle(world, provider.metrics, partialTicks, station);
+		float siderealAngle = (float)SolarSystem.calculateSiderealAngle(world, partialTicks, station.orbiting);
 		float celestialPhase = (1 - (solarAngle + 0.5F) % 1) * 2 - 1;
 
 		float starBrightness = world.getStarBrightness(partialTicks);
 
-		renderStars(partialTicks, world, mc, starBrightness, solarAngle, orbitalTilt);
+		renderStars(partialTicks, world, mc, starBrightness, solarAngle + siderealAngle, orbitalTilt);
 
 		GL11.glPushMatrix();
 		{
@@ -57,9 +60,9 @@ public class SkyProviderOrbit extends SkyProviderCelestial {
 
 			OpenGlHelper.glBlendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE, GL11.GL_ONE, GL11.GL_ZERO);
 
-			double sunSize = SolarSystem.calculateSunSize(station.orbiting);
+			double sunSize = SolarSystem.calculateSunSize(station.orbiting) * SolarSystem.SUN_RENDER_SCALE;
 			if(station.state != StationState.ORBIT) {
-				double sunTargetSize = SolarSystem.calculateSunSize(station.target);
+				double sunTargetSize = SolarSystem.calculateSunSize(station.target) * SolarSystem.SUN_RENDER_SCALE;
 				sunSize = BobMathUtil.lerp(progress, sunSize, sunTargetSize);
 			}
 			double coronaSize = sunSize * (3 - Library.smoothstep(Math.abs(celestialPhase), 0.7, 0.8));
@@ -68,19 +71,9 @@ public class SkyProviderOrbit extends SkyProviderCelestial {
 			renderSun(partialTicks, world, mc, station.orbiting.getStar(), sunSize, coronaSize, 1, 0);
 
 			CelestialBody orbiting = station.orbiting;
-			List<AstroMetric> metrics;
-			if(station.state == StationState.ORBIT) {
-				double altitude = provider.getOrbitalAltitude(station.orbiting);
-				metrics = SolarSystem.calculateMetricsFromSatellite(world, partialTicks, station.orbiting, altitude);
-			} else {
-				double fromAlt = provider.getOrbitalAltitude(station.orbiting);
-				double toAlt = provider.getOrbitalAltitude(station.target);
-				metrics = SolarSystem.calculateMetricsBetweenSatelliteOrbits(world, partialTicks, station.orbiting, station.target, fromAlt, toAlt, progress);
+			if(station.state != StationState.ORBIT && progress > 0.5) orbiting = station.target;
 
-				if(progress > 0.5) orbiting = station.target;
-			}
-
-			renderCelestials(partialTicks, world, mc, metrics, solarAngle, null, Vec3.createVectorHelper(0, 0, 0), 1, 1, orbiting, 160);
+			renderCelestials(partialTicks, world, mc, provider.metrics, solarAngle, null, Vec3.createVectorHelper(0, 0, 0), 1, 1, orbiting, SolarSystem.MAX_APPARENT_SIZE_ORBIT);
 
 		}
 		GL11.glPopMatrix();
@@ -103,7 +96,7 @@ public class SkyProviderOrbit extends SkyProviderCelestial {
 	}
 
 	// All angles within are normalized to -180/180
-	private float getCelestialAngle(WorldClient world, float partialTicks, OrbitalStation station) {
+	private float getCelestialAngle(WorldClient world, List<AstroMetric> metrics, float partialTicks, OrbitalStation station) {
 		float solarAngle = world.getCelestialAngle(partialTicks);
 		if(station.state == StationState.ORBIT) return solarAngle;
 
@@ -112,7 +105,7 @@ public class SkyProviderOrbit extends SkyProviderCelestial {
 		if(station.state != StationState.ARRIVING) lastBody = station.orbiting;
 
 		double progress = station.getUnscaledProgress(partialTicks);
-		float travelAngle = -(float)SolarSystem.calculateSingleAngle(world, partialTicks, lastBody, station.target);
+		float travelAngle = -(float)SolarSystem.calculateSingleAngle(metrics, lastBody, station.target);
 		travelAngle = MathHelper.wrapAngleTo180_float(travelAngle + 90.0F);
 
 		if(station.state == StationState.TRANSFER) {